Split AbstractDispatcherRegistry and its interface.
